step 7

+++ step7/src/main/java/io/vertx/wiki/http/HttpServerVerticle.java
@@ -10,8 +10,13 @@ import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
+import io.vertx.ext.auth.KeyStoreOptions;
+import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.jdbc.JDBCAuth;
+import io.vertx.ext.auth.jwt.JWTAuth;
+import io.vertx.ext.auth.jwt.JWTAuthOptions;
 import io.vertx.ext.jdbc.JDBCClient;
+import io.vertx.ext.jwt.JWTOptions;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.client.HttpResponse;
@@ -89,6 +94,14 @@ public class HttpServerVerticle extends AbstractVerticle {
     });
 
     Router apiRouter = Router.router(vertx);
+
+    JWTAuth jwtAuth = JWTAuth.create(vertx, new JWTAuthOptions()
+      .setKeyStore(new KeyStoreOptions()
+        .setPath("keystore.jceks")
+        .setType("jceks")
+        .setPassword("secret")));
+    apiRouter.route().handler(JWTAuthHandler.create(jwtAuth, "/api/token"));
+
     apiRouter.get("/pages").handler(this::apiRoot);
     apiRouter.get("/pages/:id").handler(this::apiGetPage);
     apiRouter.post().handler(BodyHandler.create());
@@ -96,6 +109,34 @@ public class HttpServerVerticle extends AbstractVerticle {
     apiRouter.put().handler(BodyHandler.create());
     apiRouter.put("/pages/:id").handler(this::apiUpdatePage);
     apiRouter.delete("/pages/:id").handler(this::apiDeletePage);
+    apiRouter.get("/token").handler(context -> {
+      JsonObject creds = new JsonObject()
+        .put("username", context.request().getHeader("login"))
+        .put("password", context.request().getHeader("password"));
+      auth.authenticate(creds, authResult -> {
+        if (authResult.succeeded()) {
+          User user = authResult.result();
+          user.isAuthorized("create", canCreate -> {
+            user.isAuthorized("delete", canDelete -> {
+              user.isAuthorized("update", canUpdate -> {
+                String token = jwtAuth.generateToken(
+                  new JsonObject()
+                    .put("username", context.request().getHeader("login"))
+                    .put("canCreate", canCreate.succeeded() & canCreate.result())
+                    .put("canDelete", canDelete.succeeded() & canDelete.result())
+                    .put("canUpdate", canUpdate.succeeded() & canUpdate.result()),
+                  new JWTOptions()
+                    .setSubject("Wiki API")
+                    .setIssuer("Vert.x"));
+                context.response().putHeader("Content-Type", "text/plain").end(token);
+              });
+            });
+          });
+        } else {
+          context.fail(401);
+        }
+      });
+    });
     router.mountSubRouter("/api", apiRouter);
 
     templateEngine = FreeMarkerTemplateEngine.create(vertx);
@@ -364,10 +405,14 @@ public class HttpServerVerticle extends AbstractVerticle {
   }
 
   private void apiDeletePage(RoutingContext context) {
-    int id = Integer.valueOf(context.request().getParam("id"));
-    dbService.deletePage(id, reply -> {
-      handleSimpleDbReply(context, reply);
-    });
+    if (context.user().principal().getBoolean("canDelete", false)) {
+      int id = Integer.valueOf(context.request().getParam("id"));
+      dbService.deletePage(id, reply -> {
+        handleSimpleDbReply(context, reply);
+      });
+    } else {
+      context.fail(401);
+    }
   }
 
   private boolean validateJsonPageDocument(RoutingContext context, JsonObject page, String... expectedKeys) {
